+++ messages/en.json
@@ -136,7 +136,17 @@
     "삭제한 댓글이에요": "This comment has been deleted by you.",
     "삭제된 댓글이에요": "This comment has been deleted.",
     "정말 삭제하시겠어요?": "Are you sure you want to delete this?",
-    "삭제한 한줄리뷰는 다시 복구할 수 없어요": "Once you delete a one-line review,\nit cannot be restored."
+    "삭제한 한줄리뷰는 다시 복구할 수 없어요": "Once you delete a one-line review,\nit cannot be restored.",
+    "리뷰 평가 기능은 로그인 후 이용할 수 있어요": "The review rating feature is available after logging in.",
+    "로그인 후 다시 이용해 주세요": "Please log in and try again.",
+    "리뷰 평가에 실패했어요": "Failed to rate the review.",
+    "다시 시도해 주세요": "Please try again.",
+    "댓글을 불러오는데 문제가 발생했어요 다시 시도해 주세요": "An error occurred while loading comments.\nPlease try again.",
+    "첫 댓글을 작성해 주세요": "Please write the first comment.",
+    "댓글 기능은 로그인 후 이용할 수 있어요": "The comment feature is available after logging in.",
+    "댓글 작성에 실패했어요": "Failed to post the comment.",
+    "댓글이 정상적으로 삭제되었어요": "The comment has been successfully deleted.",
+    "댓글 삭제에 실패했어요": "Failed to delete the comment."
   },
   "follow_page": {
     "인플루언서": "Influencer",
