+++ messages/ko.json
@@ -136,7 +136,17 @@
     "삭제한 댓글이에요": "삭제한 댓글이에요.",
     "삭제된 댓글이에요": "삭제된 댓글이에요.",
     "정말 삭제하시겠어요?": "정말 삭제하시겠어요?",
-    "삭제한 한줄리뷰는 다시 복구할 수 없어요": "삭제한 한줄리뷰는\n다시 복구할 수 없어요."
+    "삭제한 한줄리뷰는 다시 복구할 수 없어요": "삭제한 한줄리뷰는\n다시 복구할 수 없어요.",
+    "리뷰 평가 기능은 로그인 후 이용할 수 있어요": "리뷰 평가 기능은 로그인 후 이용할 수 있어요.",
+    "로그인 후 다시 이용해 주세요": "로그인 후 다시 이용해 주세요.",
+    "리뷰 평가에 실패했어요": "리뷰 평가에 실패했어요.",
+    "다시 시도해 주세요": "다시 시도해 주세요.",
+    "댓글을 불러오는데 문제가 발생했어요 다시 시도해 주세요": "댓글을 불러오는데 문제가 발생했어요.\n다시 시도해 주세요.",
+    "첫 댓글을 작성해 주세요": "첫 댓글을 작성해 주세요.",
+    "댓글 기능은 로그인 후 이용할 수 있어요": "댓글 기능은 로그인 후 이용할 수 있어요.",
+    "댓글 작성에 실패했어요": "댓글 작성에 실패했어요.",
+    "댓글이 정상적으로 삭제되었어요": "댓글이 정상적으로 삭제되었어요.",
+    "댓글 삭제에 실패했어요": "댓글 작성에 실패했어요."
   },
   "follow_page": {
     "인플루언서": "인플루언서",
